restructure for packaging
