+++ gclr/exec.py
@@ -1,0 +1,11 @@
+import test
+import cfg
+
+def main():
+    """ The main function of the script
+        This function run as a starting point of the program
+    """
+    #clone assignments from GitHub
+	cfg.get_assignement();
+	#export result
+	test.export_result();
